requirements.txt
faiss_cpu==1.9.0.post1
Faker==37.1.0
fitz==0.0.1.dev2
numpy==2.2.4
ollama==0.4.7
opencv_python==4.11.0.86
pypinyin==0.53.0
python_docx==1.1.2
Requests==2.32.3
setuptools==72.1.0
tqdm==4.66.4
